refactor delete review

+++ server/services/review.ts
@@ -60,11 +60,11 @@ export const ReviewService: IReviewService = {
       throw Error('can not get review frmo db');
     }
   },
-  async deleteReview(isbn: number): Promise<void> {
+  async deleteReview(isbn: number, userId: string): Promise<void> {
     try {
       const book = await BookModel.getByIsbn(isbn);
-      await ReviewModel.findOneAndDelete({ book: book._id });
-    } catch {
+      await ReviewModel.findOneAndDelete({ book: book._id, reviewer: userId });
+    } catch (err) {
       throw Error('can not get review frmo db');
     }
   },
